backend/requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.5

# Database
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
cryptography>=40.0.0
python-dotenv>=1.0.0

# CORS & Middleware
fastapi-cors>=0.0.6

# Data Export
openpyxl>=3.0.0
python-dateutil>=2.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.24.0

# Environment
python-dotenv==1.0.0
